Ready for Nick to work on PreciousStones API Implementation, see changes...
* Added more .class files to the gitignore
* Added descriptions of what it takes to get ShieldPluginManager
methods to return true or false
* Finished ShieldPM for all API methods
* Added Protect interface, which extends the ShieldAPI interface. This
is now implemented by every Protect_Class to make sure they have all
the required methods. Methods that we want to be in the visible API go
in ShieldAPI, methods we need every class to have but are NOT part of
the API (isEnabled, getPluginName, etc.) go in the Protect interface.
* Added a blank Protect_Class template as a .txt file. Might be useful
for us when we're adding more supported plugins, or if someone else
decides to write methods for a plugin they want supported.
* Added both a PreciousStones.jar and a Regios.jar to the Library
* Templated classes for both PreciousStones and Regios are completely
ready to go.
* Changed the plugin loading messages a bit. Now says "Detected
WorldGuard: (Hooked, or Loaded)", and doesn't repeat the message from
the WorldGuard class if the plugin was already loaded. Looks alot
nicer, i think.