Plugin manager is fixed for good now. i think...
I added a new method so that rather then checking if the args types are
equal, we can check if they are equal, or if the extend each other. I
mean, the way I was checking before, Player and Entity would return
false, even tho the code would have worked. Now, you can pass in a
player to a method that takes an Entity as its param and it passes the
check.

Also added more debug to figure out that that was the issue, commenting
it out for now.