Added ShieldGroup...
* Added ShieldGroup object. This is a class to store a hashSet of
player names. It has built in methods for adding and removing players
either by object or name and in lists and sets. This object is now
saved and returned for flags and any other api calls that require a
group of players to be returned.
* Updated FlagManager and Flag classes to use the new ShieldGroups
* Added a new createFlag method with a ShieldGroup param, as well as
updating the others to now make shieldGroups as well
* Added getOwners() to Protect interface
* Finished getOwners() in Regios, Residence, and WorldGuard
* Added boolean isOwner(Player) to the ShieldRegion class
* Added ShieldGroup getOwners() to ShieldRegion